Exception handling - delete

+++ src/main/java/com/gabkt/webservice/services/UserService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import com.gabkt.webservice.entities.User;
 import com.gabkt.webservice.repositories.UserRepository;
+import com.gabkt.webservice.services.exceptions.DatabaseException;
 import com.gabkt.webservice.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -19,26 +20,29 @@ public class UserService {
 	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
-		return obj.orElseThrow(()-> new ResourceNotFoundException(id));
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new DatabaseException("Database error, resource not found");
+		}
 		repository.deleteById(id);
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
 		return repository.save(entity);
 	}
-	
+
 	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
